feat(lmc): add main pilot module source code.
The lmc module implements the main function of the emulator.

However this module interferes with the Sccroll unit testing module
functionality, and thus cannot be included during the linking. The
Makefile recipes have thus been modified to include the lmc module only
during the main program compilation, and not during units testing.

Signed-off-by: Alexandre Martos <contact@amartos.fr>