fix(computer): standardize hexadecimal output values
The output format for memory values and addresses is now defined in one
place, in the computer module header.

The input format for command line has been simplified ; as the command
line expects a newline as input terminator, there is no need to limit
the string length at first. An error will be raised if the input value
is not of the expected length, however.

Signed-off-by: Alexandre Martos <contact@amartos.fr>