fix: wrong indirection value used
Since the creation of the lmc_indirection function, the indirection type
was recalculated using the opcode. However, this was not taking account
of the ADD/SUB/NAND opcodes, which modified the current opcode *before*
the indirection was done. Thus, the indirection was recalculated using a
wrong opcode, and thus a potentially wrong indirection type in those
cases, hence an error in calculations.

Ref: db4f8c554c
Signed-off-by: Alexandre Martos <contact@amartos.fr>