feat(lexer): Reduce yyparse number of parameters
The LmcLexer structure now contains all necessary informations for the
yyparse function and related. The structure is extensible, which fixes
future growing number of parameters for these.

The LmcMemoryArray structure has been created in order to keep track of
the values analyzed.

As all the values are now contained in one structure, this also reduces
the total number of fwrite calls of the compiler -- which now only
writes once after the translation.

Signed-off-by: Alexandre Martos <contact@amartos.fr>