refactor(computer): Move basic specs to their own header file
The lexer dependency for the computer module header was becoming
problematic for implementation of new features depending on the lexer.
This obviously would introduce some circular dependencies, thus
impossible.

The specs of the LMC computer have now been moved to their own header,
becoming a common dependency for all modules.

Signed-off-by: Alexandre Martos <contact@amartos.fr>