fix(computer): LmcMemoryArray not in the right thematic module
LmcMemoryArray is used in the lexer and all modules depending on it.
However, it was defined in the computer module header.

The structure has thus been moved. It also has been renamed to clarify
its properties and avoid confusion: it is not linked to the LmcMemory
structure, as the former name suggested.

Signed-off-by: Alexandre Martos <contact@amartos.fr>