refactor(computer): split operations handling in sub-functions
The one-big-switch operations function was hard to maintain and
understand, thus it has been split into several functions. This also
allows to use parts of the operations definition.

The duplicated code for with/without opcodes was then merged into the
functions, simplifying its maintenance.

This causes the lmc_phase* functions definitions to be independant of
the _UCODES macro.

Signed-off-by: Alexandre Martos <contact@amartos.fr>